# api/docs/swagger.yaml
definitions:
...
  auth_service.ClientPlatform:
    properties:
      id:
        type: string
      name:
        type: string
      project_id:
        type: string
      subdomain:
        type: string
    type: object
...
  auth_service.ClientType:
    properties:
      confirm_by:
        type: integer
      id:
        type: string
      name:
        type: string
      project_id:
        type: string
      self_recover:
        type: boolean
      self_register:
        type: boolean
    type: object
...
  auth_service.HasAccessRequest:
    properties:
      access_token:
        type: string
      client_platform_id:
        type: string
      method:
        type: string
      path:
        type: string
      project_id:
        type: string
    type: object
  auth_service.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  auth_service.LoginResponse:
    properties:
      client_platform:
        $ref: '#/definitions/auth_service.ClientPlatform'
      client_type:
        $ref: '#/definitions/auth_service.ClientType'
      permissions:
        items:
          $ref: '#/definitions/auth_service.Permission'
        type: array
      role:
        $ref: '#/definitions/auth_service.Role'
      sessions:
        items:
          $ref: '#/definitions/auth_service.Session'
        type: array
      token:
        $ref: '#/definitions/auth_service.Token'
      user:
        $ref: '#/definitions/auth_service.User'
      user_found:
        type: boolean
    type: object
  auth_service.LogoutRequest:
    properties:
      access_token:
        type: string
    type: object
  auth_service.Permission:
    properties:
      client_platform_id:
        type: string
      id:
        type: string
      name:
        type: string
      parent_id:
        type: string
    type: object
...
  auth_service.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    type: object
...
  auth_service.Role:
    properties:
      client_platform_id:
        type: string
      client_type_id:
        type: string
      id:
        type: string
      name:
        type: string
      project_id:
        type: string
    type: object
...
  auth_service.Session:
    properties:
      client_platform_id:
        type: string
      client_type_id:
        type: string
      created_at:
        type: string
      data:
        type: string
      expires_at:
        type: string
      id:
        type: string
      integration_id:
        type: string
      ip:
        type: string
      project_id:
        type: string
      role_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  auth_service.Token:
    properties:
      access_token:
        type: string
      created_at:
        type: string
      expires_at:
        type: string
      refresh_in_seconds:
        type: integer
      refresh_token:
        type: string
      updated_at:
        type: string
    type: object
...
  auth_service.User:
    properties:
      active:
        type: integer
      client_platform_id:
        type: string
      client_type_id:
        type: string
      created_at:
        type: string
      email:
        type: string
      expires_at:
        type: string
      id:
        type: string
      login:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      photo_url:
        type: string
      project_id:
        type: string
      role_id:
        type: string
      updated_at:
        type: string
    type: object
...
  http.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: string
    type: object
info:
  contact: {}
  description: This is a api gateway
  termsOfService: https://udevs.io
paths:
...
  /has-access:
    post:
      consumes:
      - application/json
      description: Has Access
      operationId: has_access
      parameters:
      - description: HasAccessRequestBody
        in: body
        name: has-access
        required: true
        schema:
          $ref: '#/definitions/auth_service.HasAccessRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.User'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Has Access
      tags:
      - Session
  /login:
    post:
      consumes:
      - application/json
      description: Login
      operationId: login
      parameters:
      - description: LoginRequestBody
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/auth_service.LoginRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Login
      tags:
      - Session
  /logout:
    delete:
      consumes:
      - application/json
      description: Logout User
      operationId: logout
      parameters:
      - description: LogoutRequest
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/auth_service.LogoutRequest'
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Logout User
      tags:
      - Session
...
  /refresh:
    put:
      consumes:
      - application/json
      description: Refresh Token
      operationId: refresh
      parameters:
      - description: RefreshTokenRequestBody
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/auth_service.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.User'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Refresh Token
      tags:
      - Session
...
swagger: "2.0"